lua callbacks are now object bound
lua callbacks should should not be saved in some global variable but be object bound.
When creating numerous lua objects in a row which reference numerous functions this may lead to huge memory usages.
this has been changed so that the callbacks are objectbound and will be deleted after unsetting the lua object